refactor: Improve peak list handling and update analysis comparison parsing in update_datamodel
