+++ app/usecases/containers/update_datamodel.rb
@@ -70,7 +70,9 @@ module Usecases
             # Update container
             next unless (container = Container.find_by(id: child[:id]))
 
-            old_analyses_compared = parse_analyses_compared(container.extended_metadata['analyses_compared'])
+            old_analyses_compared = ComparisonDatamodelSupport.parse_analyses_compared(
+              container.extended_metadata['analyses_compared'],
+            )
 
             container.update!(
               name: child[:name],
@@ -93,14 +95,18 @@ module Usecases
             container.update!(extended_metadata: extended_metadata)
           end
 
-          next unless extended_metadata&.[]('analyses_compared').present?
+          next if extended_metadata&.[]('analyses_compared').blank?
 
-          current_analyses_compared = parse_analyses_compared(extended_metadata['analyses_compared'])
+          current_analyses_compared = ComparisonDatamodelSupport.parse_analyses_compared(
+            extended_metadata['analyses_compared'],
+          )
           next unless current_analyses_compared.is_a?(Array)
           next if current_analyses_compared == old_analyses_compared
-          next if analyses_already_in_same_dataset?(container, current_analyses_compared)
+          next if ComparisonDatamodelSupport.analyses_already_in_same_dataset?(
+            container, current_analyses_compared
+          )
 
-          generate_comparison_dataset!(container, current_analyses_compared)
+          ComparisonDatamodelSupport.generate_comparison_dataset!(container, current_analyses_compared)
         end
       end
 
@@ -151,114 +157,6 @@ module Usecases
           true
         end
       end
-
-      def parse_analyses_compared(raw)
-        case raw
-        when String
-          begin
-            JSON.parse(raw.gsub('=>', ':'))
-          rescue StandardError
-            nil
-          end
-        when Array
-          raw
-        end
-      end
-
-      def analyses_already_in_same_dataset?(container, analyses_compared)
-        file_ids = analyses_compared.map { |e| e.dig('file', 'id') }.compact
-        return false if file_ids.empty?
-
-        first_dataset_id = nil
-        file_ids.each do |fid|
-          att = Attachment.find_by(id: fid)
-          return false unless att&.attachable_type == 'Container'
-
-          parent = Container.find_by(id: att.attachable_id)
-          return false unless parent && parent.container_type == 'dataset' && parent.parent_id == container.id
-
-          if first_dataset_id.nil?
-            first_dataset_id = parent.id
-          elsif first_dataset_id != parent.id
-            return false
-          end
-        end
-
-        true
-      end
-
-      # rubocop:disable Metrics/MethodLength, Metrics/AbcSize
-      def generate_comparison_dataset!(container, analyses_compared)
-        list_file = []
-        list_file_names = []
-        combined_image_filename = 'combined_image.png'
-        created_by_user = -1
-
-        dataset_child = container.children.create(
-          name: "Comparison #{Time.now.strftime('%Y-%m-%d %H:%M:%S')}",
-          container_type: 'dataset',
-        )
-        target_id = dataset_child.id
-        new_analyses_compared_list = []
-
-        analyses_compared.each do |entry|
-          attachment = Attachment.find_by(id: entry.dig('file', 'id'))
-          next unless attachment
-
-          created_by_user = attachment.created_by
-          base = File.basename(attachment.filename, '.*')
-          base_no_compared = base.sub(/_compared_[0-9:\-]+\z/, '')
-          ts = Time.now.strftime('%Y-%m-%d-%H:%M:%S')
-          new_filename = "#{base_no_compared}_compared_#{ts}#{File.extname(attachment.filename)}"
-
-          original_path = attachment.abs_path
-          next unless File.exist?(original_path)
-
-          temp_file = Tempfile.new([base, File.extname(attachment.filename)])
-          FileUtils.cp(original_path, temp_file.path)
-
-          new_att = Attachment.new(
-            filename: new_filename,
-            created_by: created_by_user,
-            created_for: created_by_user,
-            attachable_type: 'Container',
-            attachable_id: target_id,
-            file_path: temp_file.path,
-          )
-          new_att.save!
-
-          list_file_names << new_att.filename
-          list_file << new_att.abs_path
-
-          new_entry = entry.dup
-          new_entry['file'] = { 'id' => new_att.id }
-          new_entry['dataset'] = { 'id' => target_id }
-          new_entry['analysis'] = { 'id' => target_id }
-          new_analyses_compared_list << new_entry
-        end
-
-        container.extended_metadata['analyses_compared'] = new_analyses_compared_list
-        container.update_column(:extended_metadata, container.extended_metadata) # rubocop:disable Rails/SkipsModelValidations
-
-        return if list_file.empty?
-
-        _, image = Chemotion::Jcamp::CombineImg.combine(list_file, 0, list_file_names, nil)
-        return unless image.present?
-
-        old = Attachment.find_by(filename: combined_image_filename, attachable_id: target_id)
-        old&.destroy!
-
-        Attachment.create!(
-          filename: combined_image_filename,
-          created_by: created_by_user,
-          created_for: created_by_user,
-          file_path: image.path,
-          attachable_type: 'Container',
-          attachable_id: target_id,
-          thumb: true,
-        )
-      end
-      # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
     end
   end
 end
